put ACET-specific functionality in its own manager
